+++ src/modules/initialize/initialize.js
@@ -18,8 +18,8 @@ angular.module('kalabox.initialize', [
   $urlRouterProvider.otherwise('/initialize');
 })
 .controller('InitializeCtrl',
-['$scope', '$state', 'kbox', 'globalConfig',
-  function($scope, $state, kbox, globalConfig) {
+['$scope', '$state', 'kbox', 'globalConfig', 'sites',
+  function($scope, $state, kbox, globalConfig, sites) {
 
     var gui = require('nw.gui');
     var mb = new gui.Menu({type: 'menubar'});
@@ -33,7 +33,13 @@ angular.module('kalabox.initialize', [
       if (globalConfig.provisioned) {
         // Bring engine up then navigate to dashboard.
         return kbox.then(function(kbox) {
+          // Bring engine up.
           return kbox.engine.up()
+          // Pre-load sites.
+          .then(function() {
+            return sites.get();
+          })
+        // Navigate to dashboard.
           .then(function() {
             $state.go('dashboard');
           });